CMake: Remove unnecessary wrapper script mk_rootdict.sh. Drop ROOT 5.
Requires CMake >= 3.15 (for COMMAND_EXPAND_LISTS flag and
REMOVE_DUPLICATES generator expression).

Also, add FOR_EXE flag to build_root_dictionary function. This can be
used when compiling a dictionary for inclusion in an executable, e.g.
a test.  No rootmap file will be generated, and the "library" name will
not be prefixed with "lib".