Caen1190Module: Fix incorrect error message if no global header found
This is not a fatal error or even an error at all, because it may very
well happen that an event does not contain any data for an 1190 module
(or any other module). If I understand the DAQ correctly, this would
usually occur only if such a module is defined in the crate map but not
included in the readout list; simply not having any data over threshold
would insert an empty structure for the module in the event. Either way,
module data may be missing, and that is OK, or at least not a decoding
or data format error.