THaAnalyzer: Terminate analysis after 10 consecutive read errors
This is a safety valve in case fRun->ReadEvent keeps returning
READ_ERROR instead of READ_FATAL for some reason. After all, the Rune
class could be buggy user code. If that happens, something is
probably really wrong (i.e., it's not just a single corrupted event),
and we'd better let the user figure out what's going on.