Decoder: Bugfix: Module event data were never cleared in multiblock mode
This affected certain modules that kept per-event data, which would be
incorrect for all events but the first of a block.

The per-event data clearing is currently implemented as

module->Clear("E");

which may require changes to user code in libraries.

TODO: BCI: Implement per-event data clearing in a separate virtual
function Module::ClearEvent().