CodaDecoder: Accommodate possibly misaligned 64-bit data
The problem is that CODA3 may place 64-bit data (e.g. event number,
event time) at unpredictable offsets in the uint32_t[] event buffer.
Some of these data may end up not correctly aligned on 8-byte
boundaries, depending on the alignment of the buffer start, the data
offset, and details of the DAQ configuration.

This is not an issue on x86_64. The x86_64 instruction set fully
supports misaligned loads. However, other architectures, notably
ARM/Apple Silicon, may not. This commit changes direct assignments
(mov instructions) to memcpy calls, which do not require any specific
data alignment.

It is unclear at the moment if this issue is properly guarded against
in EVIO, which contains quite a few casts from void* to uint64_t*, and
similar, without any apparent awareness of the data alignment.