CodaDecoder: Hopefully finally fix the rotten event type business
With this change, by default, the decoder reverts to reporting
event_type == 1 for physics events, as in previous analyzer versions.

The user may now set a new, experimental feature called "alternate event
type" reporting. It can be enabled with analyzer->EnableAltEvType() and
tested with analyzer->AltEvTypeEnabled(). There are identical functions
in THaEvData, which are called by THaAnalyzer. If enabled, CodaDecoder
will set event_type to tsEvType, whatever it is. IsPhysicsTrigger()
checks a new bit, kPhysicsTrigger = BIT(18), in fBits to see if this is
a physics event, even if event_type is greater than MAX_PHYS_EVTYPE
(=14). This is how this should have been implemented in the first place.

All this is *almost* binary-compatible. If a user's library derives a
decoder class from CodaDecoder, and the analysis enables AltEvType, the
IsPhysicsTrigger() method may not work as expected without recompiling
because it is inline. This seems like an unlikely scenario; I am not
aware of any custom decoders that inherit from CodaDecoder.