CMake: Support auto-downloading & building ET library
If ET is not found on the system, the source distribution is downloaded
from GitHub (currently version 16.5.0) and built locally, like EVIO.

The ET library alias has been renamed from EVIO::ET to ET::ET.

Typos in FindEVIO and evio/CMakeLists.txt.in have been corrected.

TODO: SCons support. A FindET function and an SConscript.py script
for libet are implemented, but no support yet for enabling the
build of the ET client and THaOnlRun.