CodaDecoder: Swap order of time_stamp[] and run_info in trigger bank
This is a regression introduced in commit 5c7edc41 (28-Oct-21), where
the run info word was mistakenly assumed to precede the time_stamp[]
array. Since run info is not enabled by default in CODA, this bug does
not affect any SBS and NPS data to the best of my knowledge.