MultiFileRun: Fix buggy search for init info file
If the run file name did not end with both stream and segment numbers,
the wrong init info file could be scanned. This was the case in
particular for the traditional file name pattern with just a segment
number at the end, file.dat.1.

Thanks to Sanghwa Park for reporting.